+++ src/js/number.js
@@ -61,6 +61,8 @@
 			.on(Events.keyPress, data, onKeyup);
 
 		data.$container.on( [Events.touchStart, Events.mouseDown].join(" "), Classes.arrow, data, onPointerDown);
+
+		step(data, 0);
 	}
 
 	/**
@@ -112,6 +114,27 @@
 	}
 
 	/**
+	* @method
+	* @name update
+	* @description Updates instance.
+	* @example $(".target").number("update");
+	*/
+
+	function updateInstance(data) {
+		var min = parseFloat(data.$el.attr("min")),
+			max = parseFloat(data.$el.attr("max"));
+
+		data.min  = (min || min === 0) ? min : false;
+		data.max  = (max || max === 0) ? max : false;
+		data.step = parseFloat(data.$el.attr("step")) || 1;
+		data.timer        = null;
+		data.digits       = significantDigits(data.step);
+		data.disabled     = data.$el.is(":disabled") || data.$el.is("[readonly]");
+
+		step(data, 0);
+	}
+
+	/**
 	 * @method private
 	 * @name onFocus
 	 * @description Handles instance focus
@@ -130,6 +153,8 @@
 	 */
 
 	function onBlur(e) {
+		step(e.data, 0);
+
 		e.data.$container.removeClass(RawClasses.focus);
 	}
 
@@ -233,7 +258,7 @@
 			value = data.min;
 		}
 		if (data.max !== false && value > data.max) {
-			value -= data.step;
+			value = data.max;
 		}
 
 		if (value !== oValue) {
@@ -322,7 +347,8 @@
 				// Public Methods
 
 				enable        : enable,
-				disable       : disable
+				disable       : disable,
+				update        : updateInstance
 			}
 		}),