+++ CHANGELOG.md
@@ -4,9 +4,15 @@
 
 ### 1.1.4
 
+* Added `lazy` and `lazyEdge` options to Background
+* Added `update` method to Number
+* Added `update` method to Range
+* Removed non-standard `readonly` support from Checkbox
 * Fixed accessibility issues in Checkbox
+* Fixed accessibility issues in Dropdown
 * Fixed accessibility issues in Navigation
 * Fixed accessibility issues in Number
+* Fixed accessibility issues in Pagination
 * Fixed accessibility issues in Scrollbar
 * Fixed accessibility issues in Range
 * Fixed accessibility issues in Tabs